refactor(imports): update service paths in navbar and login
